+++ showoci_data.py
@@ -156,6 +156,13 @@ class ShowOCIData(object):
         return self.service.reboot_migration_counter
 
     ##########################################################################
+    # get service reboot migration
+    ##########################################################################
+    def get_service_dbsystem_maintenance(self):
+
+        return self.service.dbsystem_maintenance
+
+    ##########################################################################
     # print print error
     ##########################################################################
     def __print_error(self, msg, e):
@@ -1410,7 +1417,11 @@ class ShowOCIData(object):
                         'lifecycle_state': instance['lifecycle_state'],
                         'console_id': instance['console_id'], 'console': instance['console'],
                         'time_created': instance['time_created'],
-                        'defined_tags': instance['defined_tags'], 'freeform_tags': instance['freeform_tags']}
+                        'agent_is_management_disabled': instance['agent_is_management_disabled'],
+                        'agent_is_monitoring_disabled': instance['agent_is_monitoring_disabled'],
+                        'defined_tags': instance['defined_tags'],
+                        'freeform_tags': instance['freeform_tags']
+                        }
 
                 # boot volumes attachments
                 boot_vol_attachement = self.service.search_multi_items(self.service.C_COMPUTE, self.service.C_COMPUTE_BOOT_VOL_ATTACH, 'instance_id', instance['id'])
@@ -1704,7 +1715,7 @@ class ShowOCIData(object):
                 value = {'name': (str(db['db_name']) + " - " + str(db['db_unique_name']) + " - " + pdb_name +
                                   str(db['db_workload']) + " - " +
                                   str(db['character_set']) + " - " + str(db['lifecycle_state']) + backupstr),
-                         'backups': self.__get_database_db_backups(db['backups']),
+                         'backups': self.__get_database_db_backups(db['backups']) if 'backups' in db else [],
                          'time_created': db['time_created'],
                          'defined_tags': db['defined_tags'],
                          'dataguard': self.__get_database_db_dataguard(db['dataguard']),
@@ -1828,7 +1839,7 @@ class ShowOCIData(object):
     #
     # class oci.database.DatabaseClient(config, **kwargs)
     #
-    # Below APIs not yet done (TBD):
+    # Below APIs not done:
     # list_db_home_patch_history_entries
     # list_db_system_patch_history_entries
     # list_data_guard_associations
@@ -1875,14 +1886,18 @@ class ShowOCIData(object):
                          'vip_ips': dbs['vip_ips'],
                          'compartment_name': dbs['compartment_name'],
                          'compartment_id': dbs['compartment_id'],
-                         'patches': self.__get_database_db_patches(dbs['patches']),
-                         'listener_port': dbs['listener_port'],
-                         'db_homes': self.__get_database_db_homes(dbs['db_homes']),
-                         'db_nodes': self.__get_database_db_nodes(dbs['db_nodes']),
                          'cluster_name': dbs['cluster_name'],
                          'time_created': dbs['time_created'],
                          'defined_tags': dbs['defined_tags'],
-                         'freeform_tags': dbs['freeform_tags']}
+                         'freeform_tags': dbs['freeform_tags'],
+                         'listener_port': dbs['listener_port'],
+                         'last_maintenance_run': dbs['last_maintenance_run'],
+                         'next_maintenance_run': dbs['next_maintenance_run'],
+                         'maintenance_window': dbs['maintenance_window'],
+                         'patches': self.__get_database_db_patches(dbs['patches']),
+                         'db_homes': self.__get_database_db_homes(dbs['db_homes']),
+                         'db_nodes': self.__get_database_db_nodes(dbs['db_nodes'])
+                         }
 
                 if dbs['data_storage_size_in_gbs']:
                     value['data'] = str(dbs['data_storage_size_in_gbs']) + "GB - " + str(dbs['data_storage_percentage']) + "%" + (" - " + dbs['storage_management'] if dbs['storage_management'] else "")
@@ -2187,7 +2202,9 @@ class ShowOCIData(object):
                          'compartment_id': bucket['compartment_id'],
                          'compartment_name': bucket['compartment_name'],
                          'region_name': bucket['region_name'],
-                         'namespace_name': bucket['namespace_name']
+                         'namespace_name': bucket['namespace_name'],
+                         'public_access_type': bucket['public_access_type'],
+                         'storage_tier': bucket['storage_tier']
                          }
 
                 value['desc'] = (