+++ showoci_service.py
@@ -23,6 +23,7 @@
 from __future__ import print_function
 import oci
 import time
+import datetime
 import os
 import platform
 
@@ -56,6 +57,7 @@ class ShowOCIFlags(object):
     read_api = False
     read_data_ai = False
     skip_identity_user_credential = False
+    skip_backups = False
 
     # is_vcn_exist_for_region
     is_vcn_exist_for_region = False
@@ -290,6 +292,7 @@ class ShowOCIService(object):
     error = 0
     warning = 0
     reboot_migration_counter = 0
+    dbsystem_maintenance = []
 
     ##########################################################################
     # Service not yet available - need to remove on availability
@@ -955,29 +958,55 @@ class ShowOCIService(object):
 
     ##########################################################################
     # Load Tenancy
+    # Password policy contributed by Josh.
     ##########################################################################
 
     def __load_identity_tenancy(self, identity, tenancy_id):
         self.__load_print_status("Tenancy")
         start_time = time.time()
-
         try:
             tenancy = identity.get_tenancy(tenancy_id).data
+
+            # Getting Tenancy Password Policy
+            password_policy = {}
             try:
-                sub_regions = identity.list_region_subscriptions(tenancy.id).data
+                password_policy_data = identity.get_authentication_policy(tenancy.id).data
+                if password_policy_data:
+                    ppd = password_policy_data.password_policy
+                    password_policy = {
+                        'is_lowercase_characters_required': str(ppd.is_lowercase_characters_required),
+                        'is_numeric_characters_required': str(ppd.is_numeric_characters_required),
+                        'is_special_characters_required': str(ppd.is_special_characters_required),
+                        'is_uppercase_characters_required': str(ppd.is_uppercase_characters_required),
+                        'is_username_containment_allowed': str(ppd.is_username_containment_allowed),
+                        'minimum_password_length': str(ppd.minimum_password_length)
+                    }
+
             except oci.exceptions.ServiceError as e:
                 if self.__check_service_error(e.code):
                     self.__load_print_auth_warning()
                 else:
                     raise
 
-            data_subs = [str(es.region_name) for es in sub_regions]
+            # Get sub regions
+            data_subs = []
+            try:
+                sub_regions = identity.list_region_subscriptions(tenancy.id).data
+                data_subs = [str(es.region_name) for es in sub_regions]
+            except oci.exceptions.ServiceError as e:
+                if self.__check_service_error(e.code):
+                    self.__load_print_auth_warning()
+                else:
+                    raise
+
+            # add the data
             data = {
                 'id': tenancy.id,
                 'name': tenancy.name,
                 'home_region_key': tenancy.home_region_key,
                 'subscribe_regions': str(', '.join(x for x in data_subs)),
-                'list_region_subscriptions': data_subs
+                'list_region_subscriptions': data_subs,
+                'password_policy': password_policy
             }
             self.data[self.C_IDENTITY][self.C_IDENTITY_TENANCY] = data
             self.__load_print_cnt(1, start_time)
@@ -1014,7 +1043,8 @@ class ShowOCIService(object):
                 all_compartments = oci.pagination.list_call_get_all_results(
                     identity.list_compartments,
                     tenancy['id'],
-                    compartment_id_in_subtree=True
+                    compartment_id_in_subtree=True,
+                    retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY
                 ).data
 
             except oci.exceptions.ServiceError as e:
@@ -1163,9 +1193,9 @@ class ShowOCIService(object):
             identity_providers = []
 
             try:
-                users = oci.pagination.list_call_get_all_results(identity.list_users, tenancy_id).data
-                groups = oci.pagination.list_call_get_all_results(identity.list_groups, tenancy_id).data
-                identity_providers = identity.list_identity_providers("SAML2", tenancy_id).data
+                users = oci.pagination.list_call_get_all_results(identity.list_users, tenancy_id, retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data
+                groups = oci.pagination.list_call_get_all_results(identity.list_groups, tenancy_id, retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data
+                identity_providers = identity.list_identity_providers("SAML2", tenancy_id, retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data
             except oci.exceptions.ServiceError as item:
                 if 'auth' in item.code.lower() or item.code == 'Forbidden':
                     self.__load_print_auth_warning()
@@ -1181,7 +1211,7 @@ class ShowOCIService(object):
                 print(".", end="")
                 try:
                     user_group_memberships = oci.pagination.list_call_get_all_results(
-                        identity.list_user_group_memberships, tenancy_id, group_id=group.id).data
+                        identity.list_user_group_memberships, tenancy_id, group_id=group.id, retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data
 
                     group_users = []
                     for ugm in user_group_memberships:
@@ -1403,7 +1433,7 @@ class ShowOCIService(object):
                     continue
 
                 try:
-                    policies = oci.pagination.list_call_get_all_results(identity.list_policies, c['id']).data
+                    policies = oci.pagination.list_call_get_all_results(identity.list_policies, c['id'], retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data
 
                     if policies:
                         datapol = []
@@ -1442,13 +1472,13 @@ class ShowOCIService(object):
             groups = self.data[self.C_IDENTITY][self.C_IDENTITY_GROUPS]
 
             try:
-                identity_providers = identity.list_identity_providers("SAML2", tenancy_id).data
+                identity_providers = identity.list_identity_providers("SAML2", tenancy_id, retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data
 
                 for d in identity_providers:
 
                     # get identity providers groups
                     try:
-                        igm = oci.pagination.list_call_get_all_results(identity.list_idp_group_mappings, d.id).data
+                        igm = oci.pagination.list_call_get_all_results(identity.list_idp_group_mappings, d.id, retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data
 
                         # get the group data
                         groupdata = []
@@ -1501,7 +1531,7 @@ class ShowOCIService(object):
         try:
             dynamic_groups = []
             try:
-                dynamic_groups = oci.pagination.list_call_get_all_results(identity.list_dynamic_groups, tenancy_id).data
+                dynamic_groups = oci.pagination.list_call_get_all_results(identity.list_dynamic_groups, tenancy_id, retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data
             except oci.exceptions.ServiceError as e:
                 if self.__check_service_error(e.code):
                     self.__load_print_auth_warning()
@@ -1538,7 +1568,7 @@ class ShowOCIService(object):
         try:
             network_sources = []
             try:
-                network_sources = oci.pagination.list_call_get_all_results(identity.list_network_sources, tenancy_id).data
+                network_sources = oci.pagination.list_call_get_all_results(identity.list_network_sources, tenancy_id, retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data
             except oci.exceptions.ServiceError as e:
                 if self.__check_service_error(e.code):
                     self.__load_print_auth_warning()
@@ -1588,7 +1618,7 @@ class ShowOCIService(object):
         try:
             tags = []
             try:
-                tags = oci.pagination.list_call_get_all_results(identity.list_cost_tracking_tags, tenancy_id).data
+                tags = oci.pagination.list_call_get_all_results(identity.list_cost_tracking_tags, tenancy_id, retry_strategy=oci.retry.DEFAULT_RETRY_STRATEGY).data
             except oci.exceptions.ServiceError as e:
                 if self.__check_service_error(e.code):
                     self.__load_print_auth_warning()
@@ -3387,9 +3417,11 @@ class ShowOCIService(object):
 
             block[self.C_BLOCK_VOLGRP] += self.__load_core_block_volume_group(block_storage, compartments)
             block[self.C_BLOCK_BOOT] += self.__load_core_block_boot(block_storage, compartments)
-            block[self.C_BLOCK_BOOTBACK] += self.__load_core_block_boot_backup(block_storage, compartments)
             block[self.C_BLOCK_VOL] += self.__load_core_block_volume(block_storage, compartments)
-            block[self.C_BLOCK_VOLBACK] += self.__load_core_block_volume_backup(block_storage, compartments)
+
+            if not self.flags.skip_backups:
+                block[self.C_BLOCK_BOOTBACK] += self.__load_core_block_boot_backup(block_storage, compartments)
+                block[self.C_BLOCK_VOLBACK] += self.__load_core_block_volume_backup(block_storage, compartments)
             print("")
 
         except oci.exceptions.RequestException:
@@ -3470,7 +3502,17 @@ class ShowOCIService(object):
                            'shape_max_vnic_attachments': 0,
                            'shape_networking_bandwidth_in_gbps': 0,
                            'shape_processor_description': "",
-                           'console_vnc_connection_string': "", 'image': "Not Found", 'image_os': "Oracle Linux"}
+                           'console_vnc_connection_string': "",
+                           'image': "Not Found",
+                           'image_os': "Oracle Linux",
+                           'agent_is_management_disabled ': "",
+                           'agent_is_monitoring_disabled': ""
+                           }
+
+                    # agent_config
+                    if arr.agent_config:
+                        val["agent_is_management_disabled"] = str(arr.agent_config.is_management_disabled)
+                        val["agent_is_monitoring_disabled"] = str(arr.agent_config.is_monitoring_disabled)
 
                     # check if vm has shape config
                     if arr.shape_config:
@@ -5640,8 +5682,75 @@ class ShowOCIService(object):
             self.__print_error("__load_database_main", e)
 
     ##########################################################################
+    # __load_database_maintatance
+    ##########################################################################
+    def __load_database_maintatance(self, database_client, maintenance_run_id, db_system_name):
+        try:
+            if not maintenance_run_id:
+                return {}
+
+            # oci.database.models.MaintenanceRun
+            mt = database_client.get_maintenance_run(maintenance_run_id).data
+            val = {'id': str(mt.id),
+                   'display_name': str(mt.display_name),
+                   'description': str(mt.description),
+                   'lifecycle_state': str(mt.lifecycle_state),
+                   'time_scheduled': str(mt.time_scheduled),
+                   'time_started': str(mt.time_started),
+                   'time_ended': str(mt.time_ended),
+                   'target_resource_type': str(mt.target_resource_type),
+                   'target_resource_id': str(mt.target_resource_id),
+                   'maintenance_type': str(mt.maintenance_type),
+                   'maintenance_subtype': str(mt.maintenance_subtype),
+                   'maintenance_display': str(mt.display_name) + " ( " + str(mt.maintenance_type) + ", " + str(mt.maintenance_subtype) + ", " + str(mt.lifecycle_state) + " ), Scheduled: " + str(mt.time_scheduled)[0:16] + ((", Execution: " + str(mt.time_started)[0:16] + " - " + str(mt.time_ended)[0:16]) if str(mt.time_started) != 'None' else ""),
+                   'maintenance_alert': ""
+                   }
+
+            # If maintenane is less than 14 days
+            if mt.time_scheduled:
+                delta = mt.time_scheduled.date() - datetime.date.today()
+                if delta.days <= 14 and delta.days >= 0 and not mt.time_started:
+                    val['maintenance_alert'] = "DBSystem Maintenance is in " + str(delta.days).ljust(2, ' ') + " days, on " + str(mt.time_scheduled)[0:16] + " for " + db_system_name
+                    self.dbsystem_maintenance.append(val['maintenance_alert'])
+            return val
+
+        except oci.exceptions.ServiceError:
+            print("m", end="")
+            return ""
+        except oci.exceptions.RequestException:
+            print("m", end="")
+            return ""
+        except Exception as e:
+            self.__print_error("__load_database_maintatance", e)
+
+    ##########################################################################
+    # __load_database_maintatance_windows
+    ##########################################################################
+
+    def __load_database_maintatance_windows(self, maintenance_window):
+        try:
+            if not maintenance_window:
+                return {}
+
+            mw = maintenance_window
+            value = {
+                'preference': str(mw.preference),
+                'months': ", ".join([x.name for x in mw.months]) if mw.months else "",
+                'weeks_of_month': ", ".join([str(x) for x in mw.weeks_of_month]) if mw.weeks_of_month else "",
+                'hours_of_day': ", ".join([str(x) for x in mw.hours_of_day]) if mw.hours_of_day else "",
+                'days_of_week': ", ".join([str(x.name) for x in mw.days_of_week]) if mw.days_of_week else "",
+                'lead_time_in_weeks': str(mw.lead_time_in_weeks) if mw.lead_time_in_weeks else "",
+            }
+            value['display'] = str(mw.preference) if str(mw.preference) == "NO_PREFERENCE" else (str(mw.preference) + ": Months: " + value['months'] + ", Weeks: " + value['weeks_of_month'] + ", DOW: " + value['days_of_week'] + ", Hours: " + value['hours_of_day'] + ", Lead Weeks: " + value['lead_time_in_weeks'])
+            return value
+
+        except Exception as e:
+            self.__print_error("__load_database_maintatance_windows", e)
+
+    ##########################################################################
     # __load_database_dbsystems
     ##########################################################################
+
     def __load_database_dbsystems(self, database_client, virtual_network, compartments):
 
         data = []
@@ -5712,6 +5821,9 @@ class ShowOCIService(object):
                              'storage_management': "",
                              'sparse_diskgroup': str(dbs.sparse_diskgroup),
                              'reco_storage_size_in_gb': str(dbs.reco_storage_size_in_gb),
+                             'last_maintenance_run': self.__load_database_maintatance(database_client, dbs.last_maintenance_run_id, str(dbs.display_name) + " - " + str(dbs.shape)),
+                             'next_maintenance_run': self.__load_database_maintatance(database_client, dbs.next_maintenance_run_id, str(dbs.display_name) + " - " + str(dbs.shape)),
+                             'maintenance_window': self.__load_database_maintatance_windows(dbs.maintenance_window),
                              'region_name': str(self.config['region']),
                              'defined_tags': [] if dbs.defined_tags is None else dbs.defined_tags,
                              'freeform_tags': [] if dbs.freeform_tags is None else dbs.freeform_tags,
@@ -5909,7 +6021,9 @@ class ShowOCIService(object):
                     if db.db_backup_config.auto_backup_enabled:
                         value['auto_backup_enabled'] = True
 
-                value['backups'] = self.__load_database_dbsystems_db_backups(database_client, db.id)
+                if not self.flags.skip_backups:
+                    value['backups'] = self.__load_database_dbsystems_db_backups(database_client, db.id)
+
                 value['dataguard'] = self.__load_database_dbsystems_db_dg(database_client, db.id)
                 data.append(value)
 
@@ -6157,7 +6271,12 @@ class ShowOCIService(object):
                              'nsg_ids': dbs.nsg_ids,
                              'private_endpoint': str(dbs.private_endpoint),
                              'private_endpoint_label': str(dbs.private_endpoint_label),
-                             'backups': self.__load_database_autonomouns_backups(database_client, dbs.id)}
+                             'backups': []
+                             }
+
+                    # load bakcups
+                    if not self.flags.skip_backups:
+                        value['backups'] = self.__load_database_autonomouns_backups(database_client, dbs.id)
 
                     # license model
                     if dbs.license_model == oci.database.models.AutonomousDatabaseSummary.LICENSE_MODEL_LICENSE_INCLUDED:
@@ -6296,7 +6415,7 @@ class ShowOCIService(object):
 
                 except oci.exceptions.ServiceError as e:
                     if self.__check_service_error(e.code):
-                        self.__load_print_auth_warning()
+                        self.__load_print_auth_warning("m", False)
                         continue
                     else:
                         raise